requirements.txt
asgiref==3.5.2
Django==4.1.1
paho-mqtt==1.6.1
Pillow==9.2.0
sqlparse==0.4.3
